Add JSON schema validation
These changes will be reverted shortly after because I do not want to keep
them, since this approach seems to produce too much work and complexity.
Still, for the case that I actually change my mind later and decide to
re-implement the validation with this approach, I can get this from the
commit history.